Added relations and modified accordingly

+++ server/services/notes.service.js
@@ -1,34 +1,28 @@
 const Notes = require("../models/Notes");
 
 const createNote = async (params) => {
-    const {title, content, userId} = params
-
+    const {title, content, userId} = params || {}
     let note = await Notes.create({
         title,
         content,
         userId
     })
-};
-
-const findByUserId = async (user) => {
-    const notes = await Notes.findAll({where: {user}})
-    return notes
-};
-
-const findNote = async (id) =>{
-    let note = await Notes.findOne({where: {id}})
     return note
 };
 
-const findAll = async (userId) => {
+const findNote = async (id, userId) =>{
+    let note = await Notes.findOne({where: {id}, userId})
+    return note
+};
+
+const findAllByUser = async (userId) => {
     const notes = await Notes.findAll({where: {userId}})
     return notes
 };
 
-const updateNote = async (id, params) => {
-    let {title, content} = params
-    const note = await Notes.findOne({where: {id}})
-    console.log(note)
+const updateNote = async (params) => {
+    let {id,userId ,title , content} = params
+    const note = await Notes.findOne({where: {id} , userId})
     await note.update({title})
     await note.update({content})
 
@@ -37,22 +31,21 @@ const updateNote = async (id, params) => {
     return note
 }
 
-const deleteNote = async (id) =>{
-    let note = await Notes.destroy({where: {id}})
+const deleteNote = async (id,userId) =>{
+    let note = await Notes.destroy({where: {id} , userId})
     return note
 };
 
-const deleteAll = async () =>{
-    let note = await Notes.truncate()
+const deleteAll = async (params) =>{
+    let note = await Notes.destroy({where: params})
     return note
 };
  
 module.exports = { 
     createNote,
     findNote,
-    findAll,
+    findAllByUser,
     updateNote,
     deleteNote,
     deleteAll,
-    findByUserId
 };
